+++ src/sentry_options.h
@@ -48,6 +48,7 @@ struct sentry_options_s {
     bool require_user_consent;
     bool symbolize_stacktraces;
     bool system_crash_reporter_enabled;
+    bool attach_screenshot;
     bool on_crash_wait_for_upload;
 
     sentry_attachment_t *attachments;