+++ tests/assertions.py
@@ -90,9 +90,9 @@ def assert_event_meta(
     }
     expected_sdk = {
         "name": "sentry.native",
-        "version": "0.8.1",
+        "version": "0.8.2",
         "packages": [
-            {"name": "github:getsentry/sentry-native", "version": "0.8.1"},
+            {"name": "github:getsentry/sentry-native", "version": "0.8.2"},
         ],
     }
     if is_android: