+++ tests/test_integration_crashpad.py
@@ -522,27 +522,35 @@ def test_disable_backend(cmake, httpserver):
 
 
 @pytest.mark.skipif(
-    sys.platform != "linux",
-    reason="on_crash_wait_for_upload option is only honored on linux",
+    sys.platform != "darwin", reason="retry mechanism test only runs on macOS"
 )
-def test_wait_for_upload(cmake, httpserver):
-    pass  # pass for now
-    # tmp_path = cmake(["sentry_example"], {"SENTRY_BACKEND": "crashpad"})
-    #
-    # # make sure we are isolated from previous runs
-    # shutil.rmtree(tmp_path / ".sentry-native", ignore_errors=True)
-    #
-    # env = dict(os.environ, SENTRY_DSN=make_dsn(httpserver))
-    # httpserver.expect_oneshot_request("/api/123456/minidump/").respond_with_data("OK")
-    #
-    # # TODO figure out a way to kill the example, but keep crashpad alive
-    # with httpserver.wait(timeout=10) as waiting:
-    #     child = run(
-    #         tmp_path, "sentry_example", ["log", "crashpad-wait-for-upload", "crash"], env=env
-    #     )
-    #     assert child.returncode  # well, it's a crash after all
-    #
-    # assert waiting.result
-    #
-    #
-    # assert len(httpserver.log) == 1
+def test_crashpad_retry(cmake, httpserver):
+    tmp_path = cmake(["sentry_example"], {"SENTRY_BACKEND": "crashpad"})
+
+    subprocess.run(
+        ["sudo", "ifconfig", "lo0", "down"]
+    )  # Disables the loopback network interface
+
+    # make sure we are isolated from previous runs
+    shutil.rmtree(tmp_path / ".sentry-native", ignore_errors=True)
+
+    env = dict(os.environ, SENTRY_DSN=make_dsn(httpserver))
+    httpserver.expect_oneshot_request("/api/123456/minidump/").respond_with_data("OK")
+
+    child = run(
+        tmp_path, "sentry_example", ["log", "crash"], env=env
+    )  # crash but fail to send data
+    assert child.returncode  # well, it's a crash after all
+
+    assert len(httpserver.log) == 0
+
+    subprocess.run(
+        ["sudo", "ifconfig", "lo0", "up"]
+    )  # Enables the loopback network interface again
+    # don't rmtree here, we don't want to be isolated (example should pick up previous crash from .sentry-native DB)
+    # we also sleep to give Crashpad enough time to handle the previous crash
+    child = run(
+        tmp_path, "sentry_example", ["log", "sleep"], env=env
+    )  # run without crashing to retry send
+
+    assert len(httpserver.log) == 1