+++ tests/unit/test_utils.c
@@ -249,8 +249,10 @@ SENTRY_TEST(os)
     TEST_CHECK(!sentry_value_is_null(os));
     TEST_CHECK(sentry_value_get_type(sentry_value_get_by_key(os, "name"))
         == SENTRY_VALUE_TYPE_STRING);
+#ifndef SENTRY_PLATFORM_NX
     TEST_CHECK(sentry_value_get_type(sentry_value_get_by_key(os, "version"))
         == SENTRY_VALUE_TYPE_STRING);
+#endif
 
     sentry_value_decref(os);
 }