+++ tests/unit/test_value.c
@@ -641,6 +641,9 @@ SENTRY_TEST(value_get_by_null_key)
 
 SENTRY_TEST(value_set_stacktrace)
 {
+#ifdef SENTRY_PLATFORM_NX
+    return SKIP_TEST();
+#endif
     sentry_value_t exc
         = sentry_value_new_exception("std::out_of_range", "vector");
     sentry_value_set_stacktrace(exc, NULL, 0);